[refactor] : AppState => initAppState라는 함수로 수정
- 이유 : filter라는 상태를 굳이 유지 할 필요가 없다고 생각됨.
