+++ src/types/domain/InitAppStateType.ts
@@ -1,0 +1,5 @@
+import { Filter } from "../../domain/Filter";
+
+export interface InitAppStateType {
+  filter: Filter;
+}
